Educational Decision Support System
Data mining in higher education is a recent research field and this area
of research is

gaining popularity because of its potentials to educational institutes.
Data Mining can be used

in educational field to enhance our understanding of learning process to
focus on identifying,

extracting and evaluating variables related to the learning process of
students. This project

will help educational managers for better estimation of student’s
performance in future

studies.